labs/01-agents/01-pmc-abstract-search/requirements.txt
boto3 
defusedxml 
httpx
strands-agents
bedrock_agentcore